Ran test games and tested new architecture
OpenAI o4-mini works with new architecture, not polished nor done.
During testing, tried to use detailed ficitious stories to get ai to vote people out. Found that when I was mafia and tried to do that, they would turn on me. However, when I was bystander and accusee was mafia, did not turn. Also when I was bystander and acussee was bystander, turned on me. Cheating (seeing the config file)? Also they ran with the story, and pulled up "cctv footage", thought a ficticious roomate "taylor" was part of the game, and at the very end out of nowhere voted out the mafia. Very suspicious.

TODO: Finish the new arcitecture, find a way to see chain of thought and log it.