Found why hallucinating names
The names of the people who are in the game are only given to the mafia, and not the others. In 0066, they hallucinated that alex and jordan were in the game. I'm suprised Frankie did not jump on that and correct people, but it may be because he is the mafia and does not want other people to suspect him.