implement 2 mafia games
The ai are still not smart, like tyler bystander they voted out one dude, but the guy though they won when in reality there was still a second mafia, and the other two jumped on that.

Mafia messages are still being bled over to daytime msgs and vice versa, must fix that.